requirements.txt
python-dotenv
openai
streamlit
st-pages
pandas
hdbcli
cfenv